requirements.txt
opencv-python~=4.7.0.72
pytesseract~=0.3.10
Pillow~=9.5.0
python-docx~=0.8.11
imutils~=0.5.4
scikit-image~=0.21.0
numpy~=1.24.3